rearrange files and add a fake test
