+++ tests/test_unit/test_load_wingdisc.py
@@ -1,0 +1,5 @@
+import pytest
+
+
+def test_load_images():
+    assert True
